+++ libs/core/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.51.3-rc.21](https://github.com/SAP/fundamental-ngx/compare/v0.51.3-rc.20...v0.51.3-rc.21) (2024-08-21)
+
+
+### Features
+
+* **core:** implement calendar multi selection option ([#12176](https://github.com/SAP/fundamental-ngx/issues/12176)) ([9a47c4b](https://github.com/SAP/fundamental-ngx/commit/9a47c4ba443863e294f10fae4eda144a74615739))
+
+
+
+
+
+## [0.51.3-rc.20](https://github.com/SAP/fundamental-ngx/compare/v0.51.3-rc.19...v0.51.3-rc.20) (2024-08-21)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
 ## [0.51.3-rc.19](https://github.com/SAP/fundamental-ngx/compare/v0.51.3-rc.18...v0.51.3-rc.19) (2024-08-21)
 
 **Note:** Version bump only for package @fundamental-ngx/core